pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grounded-claim-verifier"
version = "0.1.0"
description = "Source-grounded numeric claim verification for synthetic training data"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "Teal Maker Consulting"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = []

[project.optional-dependencies]
database = ["psycopg2-binary>=2.9.0"]
dev = ["pytest>=7.0"]

[project.scripts]
grounded-claim-verifier = "grounded_claim_verifier.cli:main"

[project.urls]
Homepage = "https://github.com/Teal-Maker/grounded-claim-verifier"

[tool.hatch.build.targets.wheel]
packages = ["src/grounded_claim_verifier"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
